// src/types/dropdown.options.ts
export interface OptionTypes {
  industry?: optionField[];
  industryGcp?: optionField[];
  countries?: optionField[];
  countriesGcp?: optionField[];
  RejectionReason?: optionField[];
  specificAWSCoSellNeeds?: optionField[];
  salesActivites?: optionField[];
  solutions?: optionField[];
  AWSProducts?: optionField[];
  useCase?: optionField[];
  stageGcp?: optionField[];
  countryStateGcp?: optionField[];
  deliveryModel?: optionField[];
  deliveryModelGcp?: optionField[];
  apnPrograms?: optionField[];
  marketingUseCase?: optionField[];
  marketingActivityChannel?: optionField[];
  competitiveTracking?: optionField[];
  stage?: optionField[];
  filterStage?: optionField[];
  privateProvider?: optionField[];
  partnerRole?: optionField[];
  authority?: optionField[];
  budget?: optionField[];
  timeline?: optionField[];
  closedLostReason?: optionField[];
  solutionArea?: optionField[];
  solutionPlay?: optionField[];
  currencyValue?: optionField[];
  procurementType?: optionField[];
  crmList?: optionField[];
  sellerCode?: optionField[];
  state?: optionField[];
  stateGcp?: optionField[];
  need?: optionField[];
  partnerEngagementType?: optionField[];
  partnerConnections?: optionField[];
  supportLevel?: optionField[];
  contractVehicle?: optionField[];
  currency?: optionField[];
  status?: optionField[];
  region?: optionField[];
  productFamily?: optionField[];
  productFamilyGcp?: optionField[];
  referralType?: optionField[];
  filterConfig?: optionField[];
  purchaseIntent?: optionField[];
  helpType?: optionField[];
  mappingList?: { label: string; value: number; crmName: string }[];
  provider?: optionField[];
  cosellProvider?: optionField[];
  stageAmp?: optionField[];
  ampClosingReason?: optionField[];
  SourceGcp?: optionField[];
}
export interface optionField {
  label: string;
  value: string;
  isPostalCodeRequired?: boolean;
  postalCodeRegex?: string;
  validationMessage?: string;
  countryCode?: string;
  isStateRequired?: boolean;
  Description?: string;
  ParentEntityValue?: string;
  crmName?: string;
  ProviderName?: string;
  IPCoSellEligible?: string;
  Name?: string;
  ID?: string;
  PublisherName?: string;
  OfferID?: number;
  AdditionalData?: string;
  MarketplaceOfferID?: string;
  MarketplaceDescription?: string;
  APIName?: any;
  OfferName?: string;
  description?: string;
  dimensionName?: string;
  apiName?: string;
  isCustomDimension?: boolean;
}